Issue #SB-13498 : Changed datatype of score field from int to double

+++ data-pipeline/assessment-aggregator/src/main/java/org/ekstep/ep/samza/domain/Aggregate.java
@@ -7,23 +7,23 @@ import java.util.List;
 
 public class Aggregate {
 
-    private int totalScore;
-    private int totalMaxScore;
+    private double totalScore;
+    private double totalMaxScore;
     private List<UDTValue> questionsList;
 
-    public Aggregate(int totalScore, int totalMaxScore, List<UDTValue> questionsList) {
+    public Aggregate(double totalScore, double totalMaxScore, List<UDTValue> questionsList) {
 
         this.totalScore = totalScore;
         this.totalMaxScore = totalMaxScore;
         this.questionsList = questionsList;
     }
 
-    public int getTotalScore() {
+    public double getTotalScore() {
         return totalScore;
     }
 
 
-    public int getTotalMaxScore() {
+    public double getTotalMaxScore() {
         return totalMaxScore;
     }
